fix: satisfy invocation owner package file-count gate
